refactor methods to PipelineInputsOutputsService
